Update style.css
There was a problem at the end of web page of the netflix clone in the "Ready to watch? Enter your email to create or restart your membership" section. 
BEFORE: whenever you type something in the input box, the typed text(email) is not visible since the color of text is black, which matches the background theme.
AFTER: changed the color to white, to make sure the typed text is visible.